requirements.txt
Flask==3.0.3
firebase-admin==6.5.0
Flask-Cors==4.0.0
python-dotenv==1.0.1
